Arreglo de la captura de mensajes de los controladores ProveedorController,	    * SubClaseArticuloController,TablaDepreciacionController,TipoActivoFijoController
