+++ swRM/bd.swrm.web/bd.swrm.web.csproj
@@ -28,11 +28,6 @@
     <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="1.0.1" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Controllers\API\" />
-    <Folder Include="Controllers\API\" />
-    <Folder Include="Controllers\API\" />
-    <Folder Include="Controllers\API\" />
-    <Folder Include="Controllers\API\" />
     <Folder Include="Controllers\MVC\" />
   </ItemGroup>
   <ItemGroup>
